refactor: centralize browser utilities
