+++ app/_client/player/lyrics.ts
