+++ app/error.tsx
@@ -2,7 +2,7 @@
 'use client';
 
 import { useEffect } from 'react';
-import { reportClientError } from '../lib/client-error-reporter';
+import { reportClientError } from './_client/client-error-reporter';
 
 export default function Error({ error, reset }: { error: Error; reset: () => void }) {
   useEffect(() => {
